package com.ticket.server altina alindi ve Feign client configurasyonu duzeltildi
