--- account-services/src/main/java/com/example/accountservices/AccountServicesApplication.java
@@ -1,13 +1,0 @@
-package com.example.accountservices;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AccountServicesApplication {
-
-    public static void main(String[] args) {
-        SpringApplication. run(AccountServicesApplication.class, args);
-    }
-
-}
