--- account-services/src/main/java/com/example/accountservices/api/AccountApi.java
@@ -1,44 +1,0 @@
-package com.example.accountservices.api;
-
-import com.example.accountservices.service.AccountService;
-import com.example.servicecommon.contract.AccountDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Slice;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.awt.print.Pageable;
-
-@RestController
-@RequestMapping("account")
-@RequiredArgsConstructor
-public class AccountApi {
-
-    private final AccountService accountService;
-
-    @GetMapping("/{id}")
-    public ResponseEntity<AccountDto> get(@PathVariable("id") String id) {
-        return ResponseEntity.ok(accountService.get(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<AccountDto> save(@RequestBody AccountDto account) {
-        return ResponseEntity.ok(accountService.save(account));
-    }
-
-    @PutMapping
-    public ResponseEntity<AccountDto> update(@PathVariable("id") String id, @RequestBody AccountDto account) {
-        return ResponseEntity.ok(accountService.update(id, account));
-    }
-
-    @DeleteMapping
-    public void delete(String id) {
-        accountService.delete(id);
-    }
-
-    @GetMapping
-    public ResponseEntity<Slice<AccountDto>> getAll(Pageable pageable) {
-        return ResponseEntity.ok(accountService.findAll(pageable));
-    }
-
-}
