--- account-services/src/main/java/com/example/accountservices/configuration/CustomBeanFactory.java
@@ -1,14 +1,0 @@
-package com.example.accountservices.configuration;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class CustomBeanFactory {
-
-    @Bean
-    public ModelMapper getModelMapper() {
-        return new ModelMapper();
-    }
-}
