--- account-services/src/main/java/com/example/accountservices/repository/AccountRepository.java
@@ -1,9 +1,0 @@
-package com.example.accountservices.repository;
-
-import com.example.accountservices.entity.Account;
-import org.springframework.data.cassandra.repository.CassandraRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AccountRepository extends CassandraRepository<Account,String> {
-}
