--- account-services/src/main/java/com/example/accountservices/service/AccountService.java
@@ -1,59 +1,0 @@
-package com.example.accountservices.service;
-
-import com.example.accountservices.entity.Account;
-import com.example.accountservices.repository.AccountRepository;
-import com.example.servicecommon.contract.AccountDto;
-import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Slice;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.Assert;
-
-import java.awt.print.Pageable;
-import java.util.Optional;
-
-@RequiredArgsConstructor
-@Service
-public class AccountService {
-
-    private final AccountRepository accountRepository;
-    private final ModelMapper modelMapper;
-
-    public AccountDto get(String id) {
-
-        Account account = accountRepository.findById(id).orElseThrow(()-> new IllegalArgumentException());
-        return modelMapper.map(account, AccountDto.class);
-
-    }
-
-    @Transactional
-    public AccountDto save(AccountDto accountDto) {
-        Account account = modelMapper.map(accountDto, Account.class);
-        accountDto.setId(account.getId());
-        return accountDto;
-    }
-    @Transactional
-    public AccountDto update(String id, AccountDto accountDto) {
-        Assert.isNull(id, "Id cannot be null");
-        Optional<Account> account = accountRepository.findById(id);
-        Account accountToUpdate = account.map(it -> {
-            it.setBirthDate(accountDto.getBirthDate());
-            it.setName(accountDto.getName());
-            it.setSurname(accountDto.getSurname());
-            return it;
-        }).orElseThrow(IllegalArgumentException::new);
-        return modelMapper.map(accountRepository.save(accountToUpdate), AccountDto.class);
-    }
-
-    @Transactional
-    public void delete(String id) {
-        Account account = accountRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException());
-        accountRepository.delete(account);
-    }
-
-    public Slice<AccountDto> findAll(Pageable pageable) {
-        return null;
-    }
-}
