+++ account-services/src/main/java/com/ticket/server/AccountServicesApplication.java
@@ -1,0 +1,13 @@
+package com.ticket.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AccountServicesApplication {
+
+    public static void main(String[] args) {
+        SpringApplication. run(AccountServicesApplication.class, args);
+    }
+
+}
