+++ account-services/src/main/java/com/ticket/server/configuration/CassandraConfiguration.java
@@ -1,0 +1,61 @@
+package com.ticket.server.configuration;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.cassandra.config.AbstractCassandraConfiguration;
+import org.springframework.data.cassandra.config.CassandraClusterFactoryBean;
+import org.springframework.data.cassandra.config.SchemaAction;
+import org.springframework.data.cassandra.repository.config.EnableCassandraRepositories;
+
+@Configuration
+@EnableCassandraRepositories
+public class CassandraConfiguration extends AbstractCassandraConfiguration {
+
+    @Value("${spcloud.cassandra.contact.point}")
+    private String contactPoint;
+
+    @Value("${spcloud.cassandra.port}")
+    private int port;
+
+    @Value("${spcloud.cassandra.keyspace.name}")
+    private String keyspaceName;
+
+    @Value("${spcloud.cassandra.username}")
+    private String username;
+
+    @Value("${spcloud.cassandra.password}")
+    private String password;
+
+    @Override
+    protected String getKeyspaceName() {
+        return keyspaceName;
+    }
+
+    @Override
+    protected int getPort() {
+        return port;
+    }
+
+    @Override
+    protected String getContactPoints() {
+        return contactPoint;
+    }
+
+    @Override
+    public SchemaAction getSchemaAction() {
+        return SchemaAction.CREATE_IF_NOT_EXISTS;
+    }
+
+    @Override
+    public String[] getEntityBasePackages() {
+        return new String[] {"com.ticketmicroservis.accountservice"};
+    }
+
+    @Override
+    public CassandraClusterFactoryBean cluster() {
+        CassandraClusterFactoryBean clusterFactoryBean = super.cluster();
+        clusterFactoryBean.setUsername(username);
+        clusterFactoryBean.setPassword(password);
+        return clusterFactoryBean;
+    }
+}
