+++ account-services/src/main/java/com/ticket/server/configuration/CustomBeanFactory.java
@@ -1,0 +1,14 @@
+package com.ticket.server.configuration;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class CustomBeanFactory {
+
+    @Bean
+    public ModelMapper getModelMapper() {
+        return new ModelMapper();
+    }
+}
