+++ account-services/src/main/java/com/ticket/server/repository/AccountRepository.java
@@ -1,0 +1,9 @@
+package com.ticket.server.repository;
+
+import com.ticket.server.entity.Account;
+import org.springframework.data.cassandra.repository.CassandraRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends CassandraRepository<Account,String> {
+}
