--- eureka-server/src/main/java/com/example/eurekaserver/EurekaserverApplication.java
@@ -1,13 +1,0 @@
-package com.example.eurekaserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaserverApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(EurekaserverApplication.class, args);
-    }
-}
