+++ eureka-server/src/main/java/com/ticket/server/EurekaserverApplication.java
@@ -1,0 +1,13 @@
+package com.ticket.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class EurekaserverApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(EurekaserverApplication.class, args);
+    }
+}
