--- notification-service/src/main/java/com/example/notificationservice/NotificationService.java
@@ -1,15 +1,0 @@
-package com.example.notificationservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class NotificationService {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NotificationService.class, args);
-    }
-
-}
