+++ notification-service/src/main/java/com/ticket/server/NotificationService.java
@@ -1,0 +1,15 @@
+package com.ticket.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class NotificationService {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NotificationService.class, args);
+    }
+
+}
