--- service-common/src/main/java/com/example/servicecommon/client/AccountServiceClient.java
@@ -1,14 +1,0 @@
-package com.example.servicecommon.client;
-
-import com.example.servicecommon.contract.AccountDto;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@FeignClient("account-service")
-public interface AccountServiceClient {
-
-    @RequestMapping("/account/{id}")
-    ResponseEntity<AccountDto> get(@PathVariable("id") String id);
-}
