--- service-common/src/main/java/com/example/servicecommon/configuration/FeignConfig.java
@@ -1,9 +1,0 @@
-package com.example.servicecommon.configuration;
-
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableFeignClients
-public class FeignConfig {
-}
