--- service-common/src/main/java/com/example/servicecommon/contract/AccountDto.java
@@ -1,30 +1,0 @@
-package com.example.servicecommon.contract;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Date;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccountDto
-{
-    private String id ;
-
-    private String username;
-
-    private String name;
-
-    private String surname;
-
-    private String email;
-
-    private Date birthDate;
-
-    public String getNameSurname() {
-        return this.name + " " + this.surname;
-    }
-
-}
