--- service-common/src/main/java/com/example/servicecommon/messaging/TicketNotification.java
@@ -1,16 +1,0 @@
-package com.example.servicecommon.messaging;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class TicketNotification {
-    private String ticketId;
-    private String accountId;
-    private String ticketDescription;
-}
