+++ service-common/src/main/java/com/ticket/server/configuration/FeignConfig.java
@@ -1,0 +1,9 @@
+package com.ticket.server.configuration;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients
+public class FeignConfig {
+}
