+++ service-common/src/main/java/com/ticket/server/contract/AccountDto.java
@@ -1,0 +1,31 @@
+package com.ticket.server.contract;
+
+import lombok.*;
+
+import java.util.Date;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+public class AccountDto
+{
+    private String id ;
+
+    private String username;
+
+    private String name;
+
+    private String surname;
+
+    private String email;
+
+    private Date birthDate;
+
+    public String getNameSurname() {
+        return this.name + " " + this.surname;
+    }
+
+}
