+++ service-common/src/main/java/com/ticket/server/messaging/TicketNotification.java
@@ -1,0 +1,16 @@
+package com.ticket.server.messaging;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class TicketNotification {
+    private String ticketId;
+    private String accountId;
+    private String ticketDescription;
+}
