--- ticket-service/src/main/java/com/example/ticketservice/TicketApplication.java
@@ -1,18 +1,0 @@
-package com.example.ticketservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@EnableEurekaClient
-@ComponentScan(basePackages = {
-        "com.example.servicecommon"
-})
-public class TicketApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TicketApplication.class, args);
-    }
-}
