--- ticket-service/src/main/java/com/example/ticketservice/api/TicketApi.java
@@ -1,38 +1,0 @@
-package com.example.ticketservice.api;
-
-import com.example.ticketservice.service.TicketService;
-import com.example.ticketservice.dto.TicketDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RequestMapping("/ticket")
-@RestController
-@RequiredArgsConstructor
-public class TicketApi {
-
-    private final TicketService ticketService;
-
-    @GetMapping("/{id}")
-    public ResponseEntity<TicketDto> getById(@PathVariable String id) {
-        return ResponseEntity.ok(ticketService.getById(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<TicketDto> createTicket(@RequestBody TicketDto ticketDto) {
-        return ResponseEntity.ok(ticketService.save(ticketDto));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<TicketDto> updateTicket(@PathVariable String id,
-                                                  @RequestBody TicketDto ticketDto) {
-        return ResponseEntity.ok(ticketService.update(id, ticketDto));
-    }
-
-    @GetMapping
-    public ResponseEntity<Page<TicketDto>> getAll(Pageable pageable) {
-        return ResponseEntity.ok(ticketService.getPagination(pageable));
-    }
-}
