--- ticket-service/src/main/java/com/example/ticketservice/configuration/TicketConfiguration.java
@@ -1,19 +1,0 @@
-package com.example.ticketservice.configuration;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaRepositories("com.example")
-@EnableElasticsearchRepositories
-public class TicketConfiguration {
-
-    @Bean
-    ModelMapper modelMapper(){
-        return new ModelMapper();
-    }
-}
