--- ticket-service/src/main/java/com/example/ticketservice/dto/TicketDto.java
@@ -1,29 +1,0 @@
-package com.example.ticketservice.dto;
-
-
-import lombok.*;
-
-import java.util.Date;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class TicketDto {
-
-    private String id;
-
-    private String description;
-
-    private String notes;
-
-    private String assignee;
-
-    private Date ticketDate;
-
-    private String priorityType;
-
-    private String ticketStatus;
-}
