--- ticket-service/src/main/java/com/example/ticketservice/entity/BaseEntity.java
@@ -1,19 +1,0 @@
-package com.example.ticketservice.entity;
-
-import org.springframework.data.annotation.CreatedDate;
-
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import java.io.Serializable;
-import java.util.Date;
-
-@MappedSuperclass
-public abstract class BaseEntity implements Serializable {
-
-    @CreatedDate
-    @Column(name = "created_at")
-    private Date cretedAt;
-
-    @Column(name = "updated_at")
-    private Date updatedAt;
-}
