--- ticket-service/src/main/java/com/example/ticketservice/entity/PriorityType.java
@@ -1,17 +1,0 @@
-package com.example.ticketservice.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum PriorityType {
-
-    URGENT("Acil"),
-    LOW("Önemsiz"),
-    HIGH("Yüksek Öncelikli");
-
-    private String label;
-
-    PriorityType(String label){
-        this.label = label;
-    }
-}
