--- ticket-service/src/main/java/com/example/ticketservice/entity/Ticket.java
@@ -1,58 +1,0 @@
-package com.example.ticketservice.entity;
-
-
-import lombok.*;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-@Table(name = "ticket")
-@EqualsAndHashCode(of = {"id"})
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Data
-public class Ticket extends BaseEntity{
-
-    @Id
-    @Getter
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
-    @Column(name = "id")
-    private String id;
-
-    @Getter
-    @Setter
-    @Column(name = "description", length = 600)
-    private String description;
-
-    @Getter
-    @Setter
-    @Column(name = "notes", length = 1000)
-    private String notes;
-
-    @Getter
-    @Setter
-    @Column(name = "assignee", length = 50)
-    private String assignee;
-
-    @Getter
-    @Setter
-    @Column(name = "ticket_date")
-    private Date ticketDate;
-
-    @Getter
-    @Setter
-    @Enumerated(EnumType.ORDINAL)
-    @Column(name = "priority_type")
-    private PriorityType priorityType;
-
-    @Getter
-    @Setter
-    @Enumerated(EnumType.ORDINAL)
-    @Column(name = "ticket_status")
-    private TicketStatus ticketStatus;
-}
