--- ticket-service/src/main/java/com/example/ticketservice/entity/TicketStatus.java
@@ -1,19 +1,0 @@
-package com.example.ticketservice.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum TicketStatus {
-
-
-    OPEN("Acik"),
-    IN_PROGRESS("Üzerinde Calisiliyor"),
-    RESOLVED("Cözüldü"),
-    CLOSED("Kapandi");
-
-    private String label;
-
-    TicketStatus(String label){
-        this.label = label;
-    }
-}
