--- ticket-service/src/main/java/com/example/ticketservice/repository/TicketRepository.java
@@ -1,7 +1,0 @@
-package com.example.ticketservice.repository;
-
-import com.example.ticketservice.entity.Ticket;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TicketRepository extends JpaRepository<Ticket,String> {
-}
