--- ticket-service/src/main/java/com/example/ticketservice/repository/elasticsearchrepo/TicketElasticRepository.java
@@ -1,7 +1,0 @@
-package com.example.ticketservice.repository.elasticsearchrepo;
-
-import com.example.ticketservice.entity.elasticsearchmodel.TicketModel;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-public interface TicketElasticRepository extends ElasticsearchRepository<TicketModel,String> {
-}
