--- ticket-service/src/main/java/com/example/ticketservice/service/TicketNotificationService.java
@@ -1,8 +1,0 @@
-package com.example.ticketservice.service;
-
-import com.example.ticketservice.entity.Ticket;
-
-public interface TicketNotificationService {
-
-    void sendToQueue(Ticket ticket);
-}
