--- ticket-service/src/main/java/com/example/ticketservice/service/TicketService.java
@@ -1,17 +1,0 @@
-package com.example.ticketservice.service;
-
-import com.example.ticketservice.dto.TicketDto;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface TicketService {
-
-    TicketDto save(TicketDto ticketDto);
-
-    TicketDto update(String id, TicketDto ticketDto);
-
-
-    TicketDto getById(String ticketId);
-
-    Page<TicketDto> getPagination(Pageable pageable);
-}
