+++ ticket-service/src/main/java/com/ticket/server/TicketApplication.java
@@ -1,0 +1,16 @@
+package com.ticket.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableEurekaClient
+@EnableFeignClients
+public class TicketApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TicketApplication.class, args);
+    }
+}
