+++ ticket-service/src/main/java/com/ticket/server/client/AccountSrvcClient.java
@@ -1,0 +1,15 @@
+package com.ticket.server.client;
+
+import com.ticket.server.contract.AccountDto;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@FeignClient("account-service")
+public interface AccountSrvcClient {
+
+    @RequestMapping("/account/{id}")
+    ResponseEntity<AccountDto> get(@PathVariable("id") String id);
+}
