+++ ticket-service/src/main/java/com/ticket/server/dto/TicketDto.java
@@ -1,0 +1,29 @@
+package com.ticket.server.dto;
+
+
+import lombok.*;
+
+import java.util.Date;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class TicketDto {
+
+    private String id;
+
+    private String description;
+
+    private String notes;
+
+    private String assignee;
+
+    private Date ticketDate;
+
+    private String priorityType;
+
+    private String ticketStatus;
+}
