+++ ticket-service/src/main/java/com/ticket/server/entity/PriorityType.java
@@ -1,0 +1,17 @@
+package com.ticket.server.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum PriorityType {
+
+    URGENT("Acil"),
+    LOW("Önemsiz"),
+    HIGH("Yüksek Öncelikli");
+
+    private String label;
+
+    PriorityType(String label){
+        this.label = label;
+    }
+}
