+++ ticket-service/src/main/java/com/ticket/server/entity/TicketStatus.java
@@ -1,0 +1,19 @@
+package com.ticket.server.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum TicketStatus {
+
+
+    OPEN("Acik"),
+    IN_PROGRESS("Üzerinde Calisiliyor"),
+    RESOLVED("Cözüldü"),
+    CLOSED("Kapandi");
+
+    private String label;
+
+    TicketStatus(String label){
+        this.label = label;
+    }
+}
