+++ ticket-service/src/main/java/com/ticket/server/repository/TicketRepository.java
@@ -1,0 +1,9 @@
+package com.ticket.server.repository;
+
+import com.ticket.server.entity.Ticket;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TicketRepository extends JpaRepository<Ticket,String> {
+}
