+++ ticket-service/src/main/java/com/ticket/server/service/TicketNotificationService.java
@@ -1,0 +1,8 @@
+package com.ticket.server.service;
+
+import com.ticket.server.entity.Ticket;
+
+public interface TicketNotificationService {
+
+    void sendToQueue(Ticket ticket);
+}
