+++ ticket-service/src/main/java/com/ticket/server/service/TicketService.java
@@ -1,0 +1,17 @@
+package com.ticket.server.service;
+
+import com.ticket.server.dto.TicketDto;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface TicketService {
+
+    TicketDto save(TicketDto ticketDto);
+
+    TicketDto update(String id, TicketDto ticketDto);
+
+
+    TicketDto getById(String ticketId);
+
+    Page<TicketDto> getPagination(Pageable pageable);
+}
