+++ ticket-service/src/main/java/com/ticket/server/service/impl/TicketNotificationServiceImpl.java
@@ -1,0 +1,29 @@
+package com.ticket.server.service.impl;
+
+import com.ticket.server.entity.Ticket;
+import com.ticket.server.messaging.TicketNotification;
+import com.ticket.server.service.TicketNotificationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.cloud.stream.annotation.EnableBinding;
+import org.springframework.cloud.stream.messaging.Source;
+import org.springframework.messaging.support.MessageBuilder;
+import org.springframework.stereotype.Service;
+
+@Service
+@EnableBinding(Source.class)                // kuyruga yazan classın anatasyonu
+@RequiredArgsConstructor
+public class TicketNotificationServiceImpl implements TicketNotificationService {
+
+    private final Source source;
+
+    @Override
+    public void sendToQueue(Ticket ticket) {
+
+        TicketNotification ticketNotification =new TicketNotification();
+        ticketNotification.setAccountId(ticket.getAssignee());
+        ticketNotification.setTicketId(ticket.getId());
+        ticketNotification.setTicketDescription(ticket.getDescription());
+
+        source.output().send(MessageBuilder.withPayload(ticketNotification).build());
+    }
+}
